# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Map App API",
    "description": "Map App Restful API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "127.0.0.1:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/map/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of map that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "summary": "get coordinates with id",
        "tags": [
          "Map"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Geo"
            }
          }
        }
      },
      "delete": {
        "summary": "delete coordinates with id",
        "tags": [
          "Map"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Geo"
            }
          }
        }
      },
      "put": {
        "summary": "update coordinates with id",
        "tags": [
          "Map"
        ],
        "description": "Update coordinates",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "type": "string",
            "description": "Update"
          },
          {
            "name": "map",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "properties": {
                "lat": { "type": "number" },
                "lng": { "type": "number" },
                "description": { "type": "string" },
                "date": { "type": "Date" }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully update"
          }
        }
      }
    },
...
  },
...
}